share the TLS certificate helper across tests
